visdrone-finetune/requirements.txt
# Dataset fetching + conversion.
huggingface_hub>=0.24.0
Pillow>=10.0.0
PyYAML>=6.0

# Inference + training.
libreyolo>=1.0.0
torch>=2.0.0
opencv-python>=4.8.0
numpy>=1.26.0
tqdm>=4.66.0
